requirements.txt
flask>=2.2
flask-cors>=3.0
werkzeug>=2.2
python-dotenv>=1.0
google-generativeai>=0.3 
google-genai>=0.6.0
flask_sqlalchemy>=3.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.10
beautifulsoup4>=4.12.0
requests>=2.31.0
youtube-transcript-api>=0.6.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pdfplumber>=0.9.0
